requirements.txt
lxml==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
peewee==3.17.8
python-dotenv==1.0.1
python-telegram-bot==21.6